requirements.txt
Flask
python-dotenv
gunicorn
openai==1.54.4
